Fixing garbled output for no trailing newline in input
Printing newline in case of eof to emulate a trailing newline.

Due to how fgetws works (reading until meeting a newline,
returning NULL, if end of stream is reached),
if there would be a trailing newline,  we never would see the eof,
so we wouldn't print double newlines.

This patch only fixes the case of a non-interactive input,
because in the interactive case, the input itself shifts the output,
if there is no new line -
example:
test⠤⡧⠄⡠⠤⡀⡠⠤⠄⠤⡧⠄
⠣⠄⠫⠭⠁⠬⠭⠂⠀⠣⠄